feat: create service , dao , application

+++ src/main/java/com/ohgiraffers/model/dto/CategoryDTO.java
@@ -16,7 +16,7 @@ public class CategoryDTO {
     this.refCategoryCode = refCategoryCode;
   }
 
-  public long getCategoryCode() {
+  public int getCategoryCode() {
     return categoryCode;
   }
